Select the browse-tables url instead of building it
