+++ frontend/src/metabase/browse/tables/TableBrowser/TableBrowserInner.tsx
@@ -7,7 +7,6 @@ import { BrowserCrumbs } from "metabase/common/components/BrowserCrumbs";
 import { Link } from "metabase/common/components/Link";
 import CS from "metabase/css/core/index.css";
 import { getUserIsAdmin } from "metabase/current-user";
-import type { DraftQuestionBuilder } from "metabase/metadata-store";
 import { getShallowDatabases as getDatabases } from "metabase/metadata-store";
 import { PLUGIN_TABLE_EDITING } from "metabase/plugins";
 import { useSelector } from "metabase/redux";
@@ -28,17 +27,11 @@ import {
 } from "../analytics";
 
 import S from "./TableBrowser.module.css";
+import { getTableUrl } from "./selectors";
 import { useDatabaseCrumb } from "./useDatabaseCrumb";
-
-type GetTableUrl = (
-  table: Table,
-  buildDraftQuestion: DraftQuestionBuilder,
-) => string;
 
 type TableBrowserProps = {
   tables: Table[];
-  getTableUrl: GetTableUrl;
-  buildDraftQuestion: DraftQuestionBuilder;
   dbId: DatabaseId;
   schemaName?: string;
   xraysEnabled?: boolean;
@@ -47,8 +40,6 @@ type TableBrowserProps = {
 
 export const TableBrowserInner = ({
   tables,
-  getTableUrl,
-  buildDraftQuestion,
   dbId,
   schemaName,
   xraysEnabled,
@@ -81,9 +72,7 @@ export const TableBrowserInner = ({
             key={table.id}
             table={table}
             dbId={dbId}
-            getTableUrl={getTableUrl}
             xraysEnabled={xraysEnabled}
-            buildDraftQuestion={buildDraftQuestion}
             canEditTables={canEditTables}
           />
         ))}
@@ -96,8 +85,6 @@ type TableBrowserItemProps = {
   table: Table;
   dbId: DatabaseId;
   xraysEnabled?: boolean;
-  buildDraftQuestion: DraftQuestionBuilder;
-  getTableUrl: GetTableUrl;
   canEditTables?: boolean;
 };
 
@@ -105,19 +92,16 @@ const TableBrowserItem = ({
   table,
   dbId,
   xraysEnabled,
-  buildDraftQuestion,
-  getTableUrl,
   canEditTables,
 }: TableBrowserItemProps) => {
+  const tableUrl = useSelector((state) => getTableUrl(state, table));
   const isVirtual = isVirtualCardId(table.id);
   const isLoading = isSyncInProgress(table);
   const isTableWritable = table.is_writable;
 
   return (
     <BrowseCard
-      to={
-        !isSyncInProgress(table) ? getTableUrl(table, buildDraftQuestion) : ""
-      }
+      to={!isSyncInProgress(table) ? tableUrl : ""}
       icon="table"
       title={table.display_name || table.name}
       // Unjustified type cast. FIXME
